EthernetTracker: save default IPConfiguration after creating it
When a new interface to be tracked does not have a IPConfiguration file stored, a
default one is created. However, the configuration was not saved neither in the data
store file nor in the dynamic memory array. This was causing a null object to be
returned when the configuration of the interface was asked from the Ethernet service.

https://jira.digi.com/browse/ADK4A-1505

Signed-off-by: David Escalona <david.escalona@digi.com>